Move stuff to plugin-custom-state.ts
