+++ src/app/ui/alphafold-superposition.tsx
@@ -8,7 +8,7 @@ import { superposeAf } from '../superposition';
 import { scaleLinear as d3ScaleLinear } from 'd3-scale';
 import { axisBottom as d3AxisBotom, axisLeft as d3AxisLeft } from 'd3-axis';
 import { select as d3Select } from 'd3-selection';
-import { PluginCustomState } from '../helpers';
+import { PluginCustomState } from '../plugin-custom-state';
 
 const _InfoIcon = <svg width='24px' height='24px' viewBox='0 0 24 24' strokeWidth='0.1px'><path fill="currentColor" d="M11,9H13V7H11M12,20C7.59,20 4,16.41 4,12C4,7.59 7.59,4 12,4C16.41,4 20,7.59 20,12C20,16.41 16.41,20 12,20M12,2A10,10 0 0,0 2,12A10,10 0 0,0 12,22A10,10 0 0,0 22,12A10,10 0 0,0 12,2M11,17H13V11H11V17Z" /></svg>;
 export function InfoIconSvg() { return _InfoIcon; }
